pyproject.toml
[tool.poetry]
name = "waste-log-api"
version = "0.1.0"
description = "Example package for a containerised waste management logging API using FastAPI, SQLModel, PostgreSQL, and nginx."
authors = ["Thomas Opdam tomopdam@gmail.com"]
package-mode = false
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.20"
pydantic = {extras = ["email"], version = "^2.11.1"}
bcrypt = "^4.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
alembic = "^1.15.2"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
